Change type of all id's to strings
The database for heimdall server uses BigInt for all of it's primary keys. A BigInt can be larger than a javascript number. Due to this, the type that comes from postgres for BigInt fields is string, not number. All of the types in the application stating that it is a number are incorrect, and are fixed in this PR.