Switch from SimpleAbac to casl
Casl is a much more maintained package and sems to suit our needs well. In order to perform this switch a large number of changes had to be made since the casl library requires full objects to be passed and not just IDs. This required a large amount of restructuring in the tests.

Additionally, this fixes a few more places where BigInts were being stored in numbers and converts them to strings.